Setup processing for the two e78 networks (bad data).
